More implementations and bug fixes
