+++ bower.json
@@ -7,6 +7,7 @@
     "angular-ui-select": "^0.19.6",
     "angular": "1.4.8",
     "bootstrap-table": "^1.11.1",
-    "froala-wysiwyg-editor": "^2.5.1"
+    "froala-wysiwyg-editor": "^2.5.1",
+    "js-cookie": "^2.1.4"
   }
 }
