+++ locales/he/gate.json
@@ -1,4 +1,5 @@
 {
-    "gate" : "שער"
+    "gate" : "שער",
+    "search": "החיפוש באמצעות מספר כרטיס / שם / אימייל"
 }
