+++ views/includes/page.jade
@@ -17,6 +17,8 @@ block wrapper
                         a( href="/#{language}/camps" )=t('camps:nav')
                     li( class="#{path[2] === 'npo' ? 'active' : undefined }" )
                         a( href="/#{language}/npo" )=t('npo:npo_name')
+                    li( class="#{path[2] === 'gate' ? 'active' : undefined }" )
+                        a( href="/#{language}/gate" )=t('gate:gate')
                     if(user.isAdmin)
                         li( class="#{path[2] === 'camps-admin' ? 'active' : undefined }" )
                             a( href="/#{language}/camps-admin" )=t('camps:nav_admin')
